Move services to new homes
